refactor: 예외 추상화 계층 추가 및 패키지 구조 변경
- Line, Section, Station 3개의 예외로 분류
