--- src/main/java/subway/exception/InvalidSectionDirectionException.java
@@ -1,7 +1,0 @@
-package subway.exception;
-
-public class InvalidSectionDirectionException extends SubwayException {
-    public InvalidSectionDirectionException(final String message) {
-        super(message);
-    }
-}
