--- src/main/java/subway/exception/StationNotFoundException.java
@@ -1,7 +1,0 @@
-package subway.exception;
-
-public class StationNotFoundException extends SubwayException {
-    public StationNotFoundException(final String message) {
-        super(message);
-    }
-}
