+++ src/main/java/subway/exception/line/LineException.java
@@ -1,0 +1,9 @@
+package subway.exception.line;
+
+import subway.exception.SubwayException;
+
+public class LineException extends SubwayException {
+    public LineException(final String message) {
+        super(message);
+    }
+}
