+++ src/main/java/subway/exception/line/LineUnconnectedException.java
@@ -1,6 +1,6 @@
-package subway.exception;
+package subway.exception.line;
 
-public class LineUnconnectedException extends SubwayException {
+public class LineUnconnectedException extends LineException {
     public LineUnconnectedException() {
         super("연결되지 않은 역이 있습니다");
     }
