+++ src/main/java/subway/exception/line/NonExistLineException.java
@@ -1,6 +1,6 @@
-package subway.exception;
+package subway.exception.line;
 
-public class NonExistLineException extends SubwayException {
+public class NonExistLineException extends LineException {
     public NonExistLineException(final String inputName) {
         super("해당 조건을 충족하는 노선은 존재하지 않습니다. (입력값 : " + inputName + ")");
     }
