+++ src/main/java/subway/exception/section/AlreadyExistSectionException.java
@@ -1,6 +1,6 @@
-package subway.exception;
+package subway.exception.section;
 
-public class AlreadyExistSectionException extends SubwayException {
+public class AlreadyExistSectionException extends SectionException {
     public AlreadyExistSectionException() {
         super("이미 존재하는 구간입니다.");
     }
