+++ src/main/java/subway/exception/section/InvalidSectionDirectionException.java
@@ -1,0 +1,7 @@
+package subway.exception.section;
+
+public class InvalidSectionDirectionException extends SectionException {
+    public InvalidSectionDirectionException(final String message) {
+        super(message);
+    }
+}
