+++ src/main/java/subway/exception/section/SectionException.java
@@ -1,0 +1,9 @@
+package subway.exception.section;
+
+import subway.exception.SubwayException;
+
+public class SectionException extends SubwayException {
+    public SectionException(final String message) {
+        super(message);
+    }
+}
