+++ src/main/java/subway/exception/station/AlreadyExistStationException.java
@@ -1,6 +1,6 @@
-package subway.exception;
+package subway.exception.station;
 
-public class AlreadyExistStationException extends SubwayException {
+public class AlreadyExistStationException extends StationException {
 
     public AlreadyExistStationException() {
         super("이미 존재하는 지하철역입니다.");
