+++ src/main/java/subway/exception/station/StationException.java
@@ -1,0 +1,9 @@
+package subway.exception.station;
+
+import subway.exception.SubwayException;
+
+public class StationException extends SubwayException {
+    public StationException(final String message) {
+        super(message);
+    }
+}
