+++ src/main/java/subway/exception/station/StationNotFoundException.java
@@ -1,0 +1,7 @@
+package subway.exception.station;
+
+public class StationNotFoundException extends StationException {
+    public StationNotFoundException(final String message) {
+        super(message);
+    }
+}
