+++ src/main/java/subway/ui/ControllerAdvice.java
@@ -5,16 +5,10 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import subway.dto.ErrorResponse;
-import subway.exception.DuplicatedNameException;
 import subway.exception.SubwayException;
 
 @RestControllerAdvice
 public class ControllerAdvice {
-
-    @ExceptionHandler
-    public ResponseEntity<ErrorResponse> handleException(DuplicatedNameException e) {
-        return ResponseEntity.badRequest().body(new ErrorResponse(e.getMessage()));
-    }
 
     @ExceptionHandler
     public ResponseEntity<ErrorResponse> handleException(SubwayException e) {
