+++ src/test/java/subway/domain/SubwayTest.java
@@ -8,8 +8,8 @@ import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
 import org.junit.jupiter.api.Test;
 import subway.exception.DuplicatedNameException;
-import subway.exception.NonExistLineException;
-import subway.exception.StationNotFoundException;
+import subway.exception.line.NonExistLineException;
+import subway.exception.station.StationNotFoundException;
 
 @DisplayNameGeneration(ReplaceUnderscores.class)
 @SuppressWarnings("NonAsciiCharacters")
